The main player now has health. If you lose all your health you die. Just like in real life.

+++ src/PlayState.as
@@ -1,5 +1,6 @@
 package  
 {
+	import flash.utils.Dictionary;
 	import org.flixel.*;
 
 	public class PlayState extends FlxState
@@ -7,17 +8,23 @@ package
 		var player:FlxSprite;
 		public var level:FlxTilemap;
 		var text:FlxText;
+		var health:FlxText;
 		var coins:FlxGroup;
 		var enemyGroup2:FlxGroup;
+		var scores:Dictionary;
 		
 		override public function create():void 
 		{
 			super.create();
 			
-			text = new FlxText(10, 10, 50, "Debug");
-			add(text);
+			scores = new Dictionary();
 			
-			coins = createEnemyGroupFor(0xffffff00,12);
+			
+			
+			
+			
+			
+			coins = createEnemyGroupFor(0xffffff00, 12);
 			enemyGroup2 = createEnemyGroupFor(0xff003366,3);
 			
 			add(coins);
@@ -28,10 +35,21 @@ package
 			player.acceleration.x = 10;
 			player.maxVelocity.x = 140;
 			player.drag.x = player.maxVelocity.x * 4;
+			player.health = 100;
 			
 			player.acceleration.y = 10;
 			player.maxVelocity.y = 80;
 			player.drag.y = player.maxVelocity.y * 4;
+			
+			text = new FlxText(10, 10, 50, "Debug");
+			add(text);
+			
+			health = new FlxText(10, 30, 50);
+			add(health);
+			updateHealth();
+
+			
+			scores[player] = 0;
 
 
 			var data:Array = new Array(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,
@@ -119,40 +137,56 @@ FlxG.bgColor = 0xffaaaaaa;
 			return group;
 		}
 		
+		private function updateHealth():void {
+				health.text = "Health: "+ player.health;
+		}
+		
+		private function updateScore():void {
+			
+		}
 		
 		private function enemyCollideWithLevel(enemy:FlxSprite, level:FlxTilemap):void {
 			if (enemy.isTouching(FlxObject.CEILING)) {
 					enemy.acceleration.y = enemy.maxVelocity.y * 4;
-					text.text = "touching ceiling";
+					
 			}
 			
 			if (enemy.isTouching(FlxObject.FLOOR)) {
 					enemy.acceleration.y = -enemy.maxVelocity.y * 4;
-					text.text = "touching floor";
+					
 			}
 			
 			if (enemy.isTouching(FlxObject.LEFT)) {
 					enemy.acceleration.x = enemy.maxVelocity.x * 4;
-					text.text = "touching left";
+					
 			}
 			
 			if (enemy.isTouching(FlxObject.RIGHT)) {
 					enemy.acceleration.x = -enemy.maxVelocity.x * 4;
-					text.text = "touching right";
+					
 			}
 		}
 		
-		private function collectCoin(enemy:FlxSprite, player:FlxSprite):void {
+		private function collectCoin(coin:FlxSprite, player:FlxSprite):void {
 			text.text = "KILL";	
-			enemy.kill();
+			coin.kill();
+			scores[player] += 1;
+			
+			updateScore();
+			
 			//	FlxG.play(SoundEffectCoin,0.3);
 			//	score++;
 			//	updateScore();
 		}
 		
 		private function die(enemy:FlxSprite, player:FlxSprite):void {
-			text.text = "YOU LOSE";
-			player.kill();
+			player.health -= 1;
+			updateHealth();
+			
+			if(player.health == 0){
+				player.kill();
+				text.text = "YOU LOSE";
+			}
 			
 			//	FlxG.play(SoundEffectCoin,0.3);
 			//	score++;
